initial package reorganization for mvp restrucuture
