--- CrowdVision/app/src/main/java/net/dividedattention/crowdvision/Constants.java
@@ -1,9 +1,0 @@
-package net.dividedattention.crowdvision;
-
-/**
- * Created by drewmahrt on 5/11/16.
- */
-public class Constants {
-    public final static String FIREBASE_BASE_URL = "https://crowdvision-d92f7.firebaseio.com/";
-    public final static String FIREBASE_EVENTS = FIREBASE_BASE_URL + "/events";
-}
