--- CrowdVision/app/src/main/java/net/dividedattention/crowdvision/adapters/EventImagesRecyclerViewAdapterOld.java
@@ -1,57 +1,0 @@
-package net.dividedattention.crowdvision.adapters;
-
-import android.content.Context;
-import android.support.v4.view.ViewCompat;
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-import android.widget.ImageView;
-
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.firebase.ui.database.FirebaseRecyclerAdapter;
-import com.google.firebase.database.DatabaseReference;
-import com.squareup.picasso.Picasso;
-
-
-import net.dividedattention.crowdvision.R;
-
-/**
- * Created by drewmahrt on 5/11/16.
- */
-public class EventImagesRecyclerViewAdapterOld extends FirebaseRecyclerAdapter<String,EventImagesRecyclerViewAdapterOld.EventViewHolder> {
-    private Context mContext;
-    private PhotoClickListener mListener;
-
-    public EventImagesRecyclerViewAdapterOld(Class<String> modelClass, int modelLayout, Class<EventViewHolder> viewHolderClass, DatabaseReference ref, Context context, PhotoClickListener listener) {
-        super(modelClass, modelLayout, viewHolderClass, ref);
-        mContext = context;
-        mListener = listener;
-    }
-
-
-    @Override
-    protected void populateViewHolder(final EventViewHolder eventViewHolder, final String s, final int i) {
-        Picasso.with(mContext)
-                .load(s)
-                .into(eventViewHolder.imageView);
-
-        ViewCompat.setTransitionName(eventViewHolder.imageView,i+"_image");
-
-        eventViewHolder.imageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mListener.onPhotoClicked(s,eventViewHolder.imageView,i,getRef(i).getKey());
-            }
-        });
-    }
-
-    public static class EventViewHolder extends RecyclerView.ViewHolder {
-        public ImageView imageView;
-
-        public EventViewHolder(View itemView) {
-            super(itemView);
-            imageView = (ImageView)itemView.findViewById(R.id.image);
-        }
-    }
-
-}
