+++ CrowdVision/app/src/main/java/net/dividedattention/crowdvision/eventphotos/PhotoClickListener.java
@@ -1,0 +1,10 @@
+package net.dividedattention.crowdvision.eventphotos;
+
+import android.widget.ImageView;
+
+/**
+ * Created by drewmahrt on 7/29/16.
+ */
+public interface PhotoClickListener{
+    void onPhotoClicked(String photoUrl, ImageView imageView, int position, String key);
+}
