--- CrowdVision/app/src/main/java/net/dividedattention/crowdvision/models/User.java
@@ -1,27 +1,0 @@
-package net.dividedattention.crowdvision.models;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Created by drewmahrt on 8/3/16.
- */
-public class User {
-    private List<String> likesList;
-
-    public User(List<String> likesList) {
-        this.likesList = likesList;
-    }
-
-    public User(){
-        likesList = new ArrayList<>();
-    }
-
-    public List<String> getLikesList() {
-        return likesList;
-    }
-
-    public void setLikesList(List<String> likesList) {
-        this.likesList = likesList;
-    }
-}
