+++ CrowdVision/app/src/main/java/net/dividedattention/crowdvision/util/AddressServiceConstants.java
@@ -1,0 +1,22 @@
+package net.dividedattention.crowdvision.util;
+
+/**
+ * Created by Drew on 8/27/16.
+ */
+
+public final class AddressServiceConstants {
+    public static final int SUCCESS_RESULT = 0;
+    public static final int FAILURE_RESULT = 1;
+    public static final String PACKAGE_NAME =
+            "ly.generalassemb.drewmahrt.placesapitest";
+    public static final String RECEIVER = PACKAGE_NAME + ".RECEIVER";
+    public static final String RESULT_DATA_CITY = PACKAGE_NAME +
+            ".RESULT_DATA_CITY";
+    public static final String RESULT_DATA_STATE = PACKAGE_NAME +
+            ".RESULT_DATA_STATE";
+    public static final String RESULT_LOC_NAME = PACKAGE_NAME +
+            ".RESULT_DATA_NAME";
+    public static final String LOCATION_DATA_EXTRA = PACKAGE_NAME +
+            ".LOCATION_DATA_EXTRA";
+
+}
