+++ CrowdVision/app/src/main/java/net/dividedattention/crowdvision/util/DynamicHeightImageView.java
@@ -1,0 +1,49 @@
+package net.dividedattention.crowdvision.util;
+
+/**
+ * Created by drewmahrt on 9/24/16.
+ */
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.widget.ImageView;
+
+/**
+ * An {@link android.widget.ImageView} layout that maintains a consistent width to height aspect ratio.
+ */
+public class DynamicHeightImageView extends ImageView {
+
+    private double mHeightRatio;
+
+    public DynamicHeightImageView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public DynamicHeightImageView(Context context) {
+        super(context);
+    }
+
+    public void setHeightRatio(double ratio) {
+        if (ratio != mHeightRatio) {
+            mHeightRatio = ratio;
+            requestLayout();
+        }
+    }
+
+    public double getHeightRatio() {
+        return mHeightRatio;
+    }
+
+    @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        if (mHeightRatio > 0.0) {
+            // set the image views size
+            int width = MeasureSpec.getSize(widthMeasureSpec);
+            int height = (int) (width * mHeightRatio);
+            setMeasuredDimension(width, height);
+        }
+        else {
+            super.onMeasure(widthMeasureSpec, heightMeasureSpec);
+        }
+    }
+}
