+++ CrowdVision/app/src/main/java/net/dividedattention/crowdvision/util/FirebaseConstants.java
@@ -1,0 +1,9 @@
+package net.dividedattention.crowdvision.util;
+
+/**
+ * Created by drewmahrt on 5/11/16.
+ */
+public class FirebaseConstants {
+    public final static String FIREBASE_BASE_URL = "https://crowdvision-d92f7.firebaseio.com/";
+    public final static String FIREBASE_EVENTS = FIREBASE_BASE_URL + "/events";
+}
