+++ CrowdVision/app/src/main/java/net/dividedattention/crowdvision/util/GlideConfiguration.java
@@ -1,0 +1,23 @@
+package net.dividedattention.crowdvision.util;
+
+import android.content.Context;
+
+import com.bumptech.glide.Glide;
+import com.bumptech.glide.GlideBuilder;
+import com.bumptech.glide.load.DecodeFormat;
+import com.bumptech.glide.module.GlideModule;
+
+/**
+ * Created by drewmahrt on 5/11/16.
+ */
+public class GlideConfiguration implements GlideModule {
+
+    @Override
+    public void applyOptions(Context context, GlideBuilder builder) {
+        builder.setDecodeFormat(DecodeFormat.PREFER_ARGB_8888);
+    }
+
+    @Override
+    public void registerComponents(Context context, Glide glide) {
+    }
+}
