+++ CrowdVision/app/src/main/res/layout/activity_create_event.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:id="@+id/coord_layout"
     android:fitsSystemWindows="true"
-    tools:context=".activities.CreateEventActivity">
+    tools:context=".eventcreate.CreateEventActivity">
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -15,7 +15,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context=".activities.CreateEventActivity"
+        tools:context=".eventcreate.CreateEventActivity"
         android:descendantFocusability="beforeDescendants"
         android:focusableInTouchMode="true">
