+++ CrowdVision/app/src/main/res/layout/activity_event_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:id="@+id/coord_layout"
     android:fitsSystemWindows="true"
-    tools:context=".activities.EventListActivity">
+    tools:context=".eventlist.EventListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
