+++ CrowdVision/app/src/main/res/layout/activity_expanded_photo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="net.dividedattention.crowdvision.activities.EventPhotosActivity">
+    tools:context="net.dividedattention.crowdvision.eventphotos.EventPhotosActivity">
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
