+++ CrowdVision/app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activities.LoginActivity">
+    tools:context=".login.LoginActivity">
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
