+++ CrowdVision/app/src/main/res/layout/content_event_list.xml
@@ -7,7 +7,7 @@
     android:paddingBottom="10dp"
     android:paddingTop="10dp"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".activities.EventListActivity"
+    tools:context=".eventlist.EventListActivity"
     tools:showIn="@layout/activity_event_list">
 
     <android.support.v4.view.ViewPager
