+++ CrowdVision/app/src/main/res/layout/content_event_photos.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".activities.EventPhotosActivity"
+    tools:context=".eventphotos.EventPhotosActivity"
     tools:showIn="@layout/fragment_all_photos">
 
     <android.support.v7.widget.RecyclerView
