+++ CrowdVision/app/src/main/res/layout/fragment_all_photos.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activities.EventPhotosActivity">
+    tools:context=".eventphotos.EventPhotosActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
