+++ CrowdVision/app/src/main/res/menu/menu_event_list.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".activities.EventListActivity">
+    tools:context=".eventlist.EventListActivity">
     <item
         android:id="@+id/action_logout"
         android:orderInCategory="1"
